Minishell final

+++ Minishell/libft/ft_atoi.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tumolabs <tumolabs@student.42.fr>          +#+  +:+       +#+        */
+/*   By: adashyan <adashyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/04/04 20:37:18 by sokhacha          #+#    #+#             */
-/*   Updated: 2023/01/24 11:55:47 by tumolabs         ###   ########.fr       */
+/*   Created: 2023/04/06 15:34:18 by adashyan          #+#    #+#             */
+/*   Updated: 2023/04/16 18:46:14 by adashyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char	*str)
+long long	ft_atoi(const char	*str)
 {
-	int	i;
-	int	j;
-	int	res;
+	int			i;
+	int			j;
+	long long	res;
 
 	i = 0;
 	res = 0;
